Moved season-hub/[event] to app directory
